+++ src/signalk/Polar.js
@@ -336,7 +336,7 @@ class Polar {
   }
 
   get trace() {
-    return Math.sqrt(this.xVariance ** 2 + this.yVariance ** 2);
+    return this.xVariance + this.yVariance;
   }
 
   report() {